updated_Faq
Updated Geodisy FAQ only. This FAQ page was recently changed completely (audit was done on previous page). I see multiple spots for ensemble/ensembles which are not covered in audit report and I'm not 100% sure on where to change preceding preposition or article when I make change from ensemble/ensembles to jeu/jeux .